adjust code after review
