+++ package.json
@@ -92,6 +92,7 @@
     "numeral": "^2.0.6",
     "pino": "^5.12.6",
     "pn": "^1.1.0",
+    "promisify-file-reader": "^4.0.0",
     "qrcode.react": "^0.9.3",
     "qs": "^6.8.0",
     "react": "^16.8.1",
